requirements.txt
gitdb==4.0.9
GitPython==3.1.24
numpy==1.21.5
pandas==1.3.5
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
smmap==5.0.0
typing-extensions==4.0.1
